401 - 500/404. Sum of Left Leaves/SolutionTest.cs
using System.Text;
using NUnit.Framework;
using System.Linq;
using System.Collections.Generic;
using Common;

namespace SumOfLeftLeaves
{
    [TestFixture]
    public class SolutionTest
    {
        private static readonly object[] testCasesStr =
        {
            new object[] {"[3,9,20,null,null,15,7]", 24},
            new object[] {"[1]", 0},
        };

        [Test]
        [TestCaseSource("testCasesStr")]
        public void Test_Generic(string treeStr, int expected)
        {
            var root = TreeNodeHelper.BuildTree(treeStr);

            var sol = new Solution();
            var res = sol.SumOfLeftLeaves(root);

            Assert.That(res, Is.EqualTo(expected));
        }
    }
}